+++ packages/interfaces/src/__test__/IAppliance.test.js
@@ -222,18 +222,6 @@ describe('IAppliance', () => {
 		})
 	})
 
-	describe('emitResult', () => {
-		it('should throw an error when called without implementation', () => {
-			const appliance = new PartiallyImplementedAppliance()
-			expect(() => appliance.emitResult()).toThrow(NotImplementedError)
-		})
-
-		it('should not throw an error when called with implementation', () => {
-			const appliance = new FullyImplementedAppliance()
-			expect(() => appliance.emitResult()).not.toThrow(NotImplementedError)
-		})
-	})
-
 	describe('duckTypeProperties', () => {
 		it('should contain all properties of an IAppliance', () => {
 			const completePropertySet = new Set([